refactoring
- changed common files to class based files
- added RandomGenerator class and new commands
